requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1

# Document Parsing
PyMuPDF==1.25.3
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2

# RAG & Embeddings
faiss-cpu>=1.12.0
sentence-transformers==3.3.1

# LLM Provider
groq==0.15.0
langchain-groq==0.1.20
langchain-core==1.2.17
google-genai==0.3.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
tenacity==9.0.0
numpy==1.26.4

# Phase 2: Video Asset Dependencies
edge-tts>=6.1.12
moviepy>=1.0.3
aiohttp>=3.9.3
